+++ sentry-spring-boot-starter/src/main/java/io/sentry/spring/boot/SentryLogbackAppenderAutoConfiguration.java
@@ -1,20 +1,13 @@
 package io.sentry.spring.boot;
 
-import ch.qos.logback.classic.Level;
-import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
-import ch.qos.logback.classic.spi.ILoggingEvent;
-import ch.qos.logback.core.Appender;
 import com.jakewharton.nopen.annotation.Open;
 import io.sentry.logback.SentryAppender;
-import java.util.Iterator;
-import java.util.Optional;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 /** Auto-configures {@link SentryAppender}. */
@@ -23,40 +16,11 @@ import org.springframework.context.annotation.Configuration;
 @ConditionalOnClass({LoggerContext.class, SentryAppender.class})
 @ConditionalOnProperty(name = "sentry.logging.enabled", havingValue = "true", matchIfMissing = true)
 @ConditionalOnBean(SentryProperties.class)
-public class SentryLogbackAppenderAutoConfiguration implements InitializingBean {
+public class SentryLogbackAppenderAutoConfiguration {
 
-  @Autowired private SentryProperties sentryProperties;
-
-  @Override
-  public void afterPropertiesSet() {
-    Logger rootLogger = (Logger) LoggerFactory.getLogger(org.slf4j.Logger.ROOT_LOGGER_NAME);
-
-    if (!isSentryAppenderRegistered(rootLogger)) {
-      SentryAppender sentryAppender = new SentryAppender();
-      sentryAppender.setName("SENTRY_APPENDER");
-      sentryAppender.setContext((LoggerContext) LoggerFactory.getILoggerFactory());
-
-      Optional.ofNullable(sentryProperties.getLogging().getMinimumBreadcrumbLevel())
-          .map(slf4jLevel -> Level.toLevel(slf4jLevel.name()))
-          .ifPresent(sentryAppender::setMinimumBreadcrumbLevel);
-      Optional.ofNullable(sentryProperties.getLogging().getMinimumEventLevel())
-          .map(slf4jLevel -> Level.toLevel(slf4jLevel.name()))
-          .ifPresent(sentryAppender::setMinimumEventLevel);
-
-      sentryAppender.start();
-      rootLogger.addAppender(sentryAppender);
-    }
-  }
-
-  private boolean isSentryAppenderRegistered(Logger logger) {
-    Iterator<Appender<ILoggingEvent>> it = logger.iteratorForAppenders();
-    while (it.hasNext()) {
-      Appender<ILoggingEvent> appender = it.next();
-
-      if (appender.getClass().equals(SentryAppender.class)) {
-        return true;
-      }
-    }
-    return false;
+  @Bean
+  public @NotNull SentryLogbackInitializer sentryLogbackInitializer(
+      final @NotNull SentryProperties sentryProperties) {
+    return new SentryLogbackInitializer(sentryProperties);
   }
 }